+++ src/components/CancelButton.vue
@@ -1,8 +1,15 @@
 <script>
 import { changeStateMixin } from '@/mixins/changeStateMixin'
+import { mapGetters } from 'vuex'
 
 export default {
-  mixins: [changeStateMixin]
+  mixins: [changeStateMixin],
+  computed: {
+    ...mapGetters('license', ['hasPermission']),
+    isReadOnly() {
+      return !this.hasPermission('update', 'run')
+    }
+  }
 }
 </script>
 
@@ -15,9 +22,7 @@ export default {
             color="red darken-3"
             class="vertical-button white--text"
             :style="{ height: '46px' }"
-            :disabled="
-              !checkVersion || role === 'READ_ONLY_USER' || isFinishing
-            "
+            :disabled="!checkVersion || isReadOnly || isFinishing"
             text
             small
             depressed
@@ -29,8 +34,8 @@ export default {
           </v-btn>
         </div>
       </template>
-      <span v-if="role === 'READ_ONLY_USER'">
-        Read-only users can't cancel flows.
+      <span v-if="isReadOnly">
+        You don't have permission to modify runs.
       </span>
       <span v-else-if="!checkVersion">
         Your Flow was registered with version {{ flowRun.flow.core_version }} of