+++ src/pages/FlowRun/Actions.vue
@@ -38,9 +38,7 @@ export default {
   },
   computed: {
     ...mapGetters('tenant', ['tenant', 'role']),
-    isReadOnlyUser() {
-      return this.role === 'READ_ONLY_USER'
-    },
+    ...mapGetters('license', ['hasPermission']),
     isScheduled() {
       return this.flowRun?.state === 'Scheduled'
     },
@@ -146,7 +144,11 @@ export default {
             text
             depressed
             :loading="runFlowNowLoading"
-            :disabled="runFlowNowLoading || runFlowNowClicked || isReadOnlyUser"
+            :disabled="
+              runFlowNowLoading ||
+                runFlowNowClicked ||
+                !hasPermission('create', 'run')
+            "
             small
             @click="runFlowNow"
           >
@@ -155,8 +157,8 @@ export default {
           </v-btn>
         </div>
       </template>
-      <span v-if="isReadOnlyUser">
-        Read-only users cannot run flows
+      <span v-if="!hasPermission('create', 'run')">
+        You don't have permission to run flows
       </span>
       <span v-else-if="runFlowNowClicked">
         This flow run has been scheduled to start as soon as possible.
@@ -175,7 +177,7 @@ export default {
             depressed
             style="height: 46px;"
             small
-            :disabled="isReadOnlyUser || !canRestart"
+            :disabled="!hasPermission('update', 'run') || !canRestart"
             color="info"
             @click="handleRestartClick"
           >
@@ -184,8 +186,8 @@ export default {
           </v-btn>
         </div>
       </template>
-      <span v-if="isReadOnlyUser">
-        Read-only users cannot restart flow runs
+      <span v-if="!hasPermission('update', 'run')">
+        You don't have permission to restart flow runs
       </span>
       <span v-else-if="!canRestart"
         >You can only restart non-archived flow runs from a failed or cancelled
@@ -232,8 +234,8 @@ export default {
           </v-btn>
         </div>
       </template>
-      <span v-if="isReadOnlyUser">
-        Read-only users cannot delete flows
+      <span v-if="!hasPermission('delete', 'run')">
+        You don't have permission to delete flows
       </span>
       <span v-else>Coming Soon!</span>
     </v-tooltip>