+++ src/pages/TeamSettings/CloudHooks.vue
@@ -78,9 +78,7 @@ export default {
   },
   computed: {
     ...mapGetters('tenant', ['tenant', 'role']),
-    isReadOnlyUser() {
-      return this.role === 'READ_ONLY_USER'
-    },
+    ...mapGetters('license', ['hasPermission']),
     isLoadingTable() {
       return this.hooksLoaded
     },
@@ -260,7 +258,7 @@ export default {
   <ManagementLayout :show="!isLoadingTable" control-show>
     <template #title>Cloud Hooks</template>
     <template #subtitle>
-      <div v-if="isReadOnlyUser">
+      <div v-if="!hasPermission('update', 'cloud-hook')">
         View your team's
         <ExternalLink
           href="https://docs.prefect.io/orchestration/concepts/cloud_hooks.html"
@@ -277,7 +275,7 @@ export default {
       </div>
     </template>
 
-    <template v-if="!isReadOnlyUser" #cta>
+    <template v-if="hasPermission('create', 'cloud-hook')" #cta>
       <v-dialog v-model="createNewCloudHook" max-width="700">
         <template #activator="{ onD }">
           <v-tooltip top>
@@ -294,7 +292,7 @@ export default {
                 </v-btn>
               </div>
             </template>
-            <span v-if="isReadOnlyUser">
+            <span v-if="!hasPermission('create', 'cloud-hook')">
               Read-only users cannot create new Cloud Hooks.
             </span>
             <span v-else>
@@ -309,7 +307,10 @@ export default {
           <v-card-text class="pl-12">
             <CloudHookForm
               v-if="createNewCloudHook"
-              :editable="!isReadOnlyUser"
+              :editable="
+                hasPermission('create', 'cloud-hook') &&
+                  hasPermission('delete', 'cloud-hook')
+              "
               edit-on-render
               @close="createNewCloudHook = false"
               @update="$apollo.queries.cloudHooks.refetch()"
@@ -441,7 +442,11 @@ export default {
                     class="v-input--vertical mt-0"
                     color="primary"
                     :loading="item.loading"
-                    :disabled="isReadOnlyUser || item.loading"
+                    :disabled="
+                      (!hasPermission('create', 'cloud-hook') &&
+                        !hasPermission('delete', 'cloud-hook')) ||
+                        item.loading
+                    "
                     @change="_handleSetCloudHookStatusChange($event, item)"
                   >
                     <template #label>
@@ -453,8 +458,8 @@ export default {
                 </div>
               </div>
             </template>
-            <span v-if="isReadOnlyUser">
-              Read-only users cannot change Cloud Hook states.
+            <span v-if="!hasPermission('update', 'cloud-hook')">
+              You don't have permission to change Cloud Hook states.
             </span>
             <span v-else>
               {{ item.active ? 'Deactivate' : 'Activate' }}
@@ -480,7 +485,13 @@ export default {
             <span class="capitalize">{{ formatStates(item.states) }}</span>
           </v-tooltip>
         </template>
-        <template v-if="!isReadOnlyUser" #item.action="{ item }">
+        <template
+          v-if="
+            hasPermission('create', 'cloud-hook') &&
+              hasPermission('delete', 'cloud-hook')
+          "
+          #item.action="{ item }"
+        >
           <v-tooltip bottom>
             <template #activator="{ on }">
               <div style="display: inline-block;" v-on="on">
@@ -562,7 +573,10 @@ export default {
           <v-card-text class="pl-12">
             <CloudHookForm
               v-if="dialogEditHook"
-              :editable="!isReadOnlyUser"
+              :editable="
+                hasPermission('create', 'cloud-hook') &&
+                  hasPermission('delete', 'cloud-hook')
+              "
               edit-on-render
               :hook="selectHook"
               show-controls