+++ src/pages/TeamSettings/FlowGroups.vue
@@ -84,6 +84,7 @@ export default {
     }
   },
   computed: {
+    ...mapGetters('license', ['hasPermission']),
     ...mapGetters('tenant', ['tenant']),
     ...mapGetters('api', ['isCloud']),
 
@@ -112,7 +113,10 @@ export default {
       return this.loadingKey > 0
     },
     isTenantAdmin() {
-      return this.tenant.role === 'TENANT_ADMIN'
+      return (
+        this.hasPermission('update', 'run') &&
+        this.hasPermission('delete', 'run')
+      )
     },
     items() {
       if (!(this.versionGroups && this.flows)) return []