+++ src/pages/TeamSettings/Secrets.vue
@@ -90,8 +90,10 @@ export default {
   },
   computed: {
     ...mapGetters('tenant', ['tenant', 'role']),
+    ...mapGetters('license', ['hasPermission']),
+
     isReadOnlyUser() {
-      return this.role === 'READ_ONLY_USER'
+      return !this.hasPermission('update', 'secret')
     },
     secretExists() {
       if (!this.secretNames) return false
@@ -320,7 +322,7 @@ export default {
         icon="lock"
         max-width="380"
       >
-        Read-only users cannot manage secrets.
+        You don't have permission to manage secrets.
       </v-alert>
     </template>