+++ src/pages/TeamSettings/TeamSettings.vue
@@ -27,7 +27,10 @@ export default {
   },
   computed: {
     ...mapGetters('api', ['isCloud']),
-    ...mapGetters('tenant', ['tenant', 'role'])
+    ...mapGetters('tenant', ['tenant', 'role']),
+    isTenantAdmin() {
+      return this.role == 'TENANT_ADMIN'
+    }
   },
   watch: {
     tenant() {
@@ -270,7 +273,7 @@ export default {
 
       <template #append>
         <v-list dense>
-          <v-list-item v-if="false && role == 'TENANT_ADMIN'" :ripple="false">
+          <v-list-item v-if="false && isTenantAdmin" :ripple="false">
             <v-list-item-content v-if="$vuetify.breakpoint.mdAndUp">
               <v-btn
                 color="red"
@@ -301,7 +304,7 @@ export default {
       </v-fade-transition>
     </div>
 
-    <template v-if="false && role == 'TENANT_ADMIN'">
+    <template v-if="false && isTenantAdmin">
       <v-dialog v-model="deleteTeamDialog" max-width="600">
         <v-card>
           <v-card-title class="text-h5 word-break-normal mb-3">
@@ -322,7 +325,7 @@ export default {
               </div>
               <v-form v-model="deleteTeamFormValid">
                 <v-text-field
-                  v-if="tenant.role == 'TENANT_ADMIN'"
+                  v-if="isTenantAdmin"
                   v-model="teamName"
                   autocomplete="off"
                   :label="tenant.name"